+++ package.json
@@ -22,23 +22,24 @@
     "validator": "^13.9.0"
   },
   "devDependencies": {
+    "@types/node": "^20.2.3",
     "@types/react": "^18.2.7",
     "@types/react-dom": "^18.2.4",
     "@types/validator": "^13.7.17",
     "@typescript-eslint/eslint-plugin": "^5.59.7",
     "@typescript-eslint/parser": "^5.59.7",
     "@vitejs/plugin-react": "^4.0.0",
-    "eslint": "^8.41.0",
-    "eslint-plugin-react": "^7.32.2",
-    "sass": "^1.62.1",
-    "typescript": "^5.0.4",
-    "vite": "^4.3.8",
     "autoprefixer": "^10.4.14",
-    "postcss": "^8.4.23",    
+    "eslint": "^8.41.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-simple-import-sort": "^10.0.0"
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "postcss": "^8.4.23",
+    "sass": "^1.62.1",
+    "typescript": "^5.0.4",
+    "vite": "^4.3.8"
   }
 }
