+++ src/shared/scss/_base.scss
@@ -1,3 +1,5 @@
+@use './utils/bootstrap' as *;
+
 html {
   font-family: 'Borna';
   font-style: normal;
