+++ src/shared/scss/_buttons.scss
@@ -1,4 +1,4 @@
-@use '/src/styles/utils/bootstrap' as *;
+@use './utils/bootstrap' as *;
 
 .action-button-container {
   display: flex;
