+++ src/shared/scss/index.scss
@@ -1,3 +1,7 @@
+@use './base';
+@use './buttons';
+@use './fonts';
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
