+++ src/types.ts
@@ -1,0 +1,4 @@
+export type PFPVideoUrl = {
+  video: string;
+  poster: string;
+};
