+++ vite.config.ts
@@ -1,7 +1,7 @@
-import react from "@vitejs/plugin-react";
+import react from '@vitejs/plugin-react';
 import autoprefixer from 'autoprefixer';
-import { defineConfig } from "vite";
-
+import * as path from 'path';
+import { defineConfig } from 'vite';
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
@@ -11,7 +11,12 @@ export default defineConfig({
   },
   css: {
     postcss: {
-      plugins: [autoprefixer]
-    }
-  }
+      plugins: [autoprefixer],
+    },
+  },
+  resolve: {
+    alias: {
+      '@scss': path.resolve(__dirname, '/src/shared/scss'),
+    },
+  },
 });
